Move dispatch to avoid setState during render
